+++ lib/pay/fake_processor/subscription.rb
@@ -29,6 +29,8 @@ module Pay
       # With trial, sets end to trial end (mimicing Stripe)
       # Without trial, sets can ends_at to end of month
       def cancel(**options)
+        return if canceled?
+
         if pay_subscription.on_trial?
           pay_subscription.update(ends_at: pay_subscription.trial_ends_at)
         else
@@ -37,6 +39,8 @@ module Pay
       end
 
       def cancel_now!(**options)
+        return if canceled?
+
         ends_at = Time.current
         pay_subscription.update(
           status: :canceled,